+++ accounts/forms.py
@@ -1,0 +1,23 @@
+from django.contrib.auth.forms import AuthenticationForm, UserCreationForm
+from django.contrib.auth.models import User
+from django import forms
+
+class LoginForm(AuthenticationForm):
+    def __init__(self, request=None, *args, **kwargs):
+        super().__init__(request, *args, **kwargs)
+
+        self.fields['username'].label = 'Name'
+        self.error_messages['invalid_login'] = 'INVALID CREDENTIALS!!! \
+        Enter correct name (first name and last name -> separate first name and last name with a space!) and password. \
+        Name and password maybe case-senstitive.'
+
+
+class SignupForm(UserCreationForm):
+    first_name = forms.CharField(required=True)
+    last_name = forms.CharField(required=True)
+    email = forms.EmailField(required=True)
+
+    class Meta:
+        model = User
+        fields = ['first_name', 'last_name', 'email', 'password1', 'password2']
+
