+++ users/forms.py
@@ -4,27 +4,6 @@ from django.contrib.auth.models import User
 from django import forms
 from .models import Voters, Aspirants, Blog
 
-
-class LoginForm(AuthenticationForm):
-    def __init__(self, request=None, *args, **kwargs):
-        super().__init__(request, *args, **kwargs)
-
-        self.fields['username'].label = 'Name'
-        self.error_messages['invalid_login'] = 'INVALID CREDENTIALS!!! \
-        Enter correct name (first name and last name -> separate first name and last name with a space!) and password. \
-        Name and password maybe case-senstitive.'
-
-
-class SignupForm(UserCreationForm):
-    first_name = forms.CharField(required=True)
-    last_name = forms.CharField(required=True)
-    email = forms.EmailField(required=True)
-
-    class Meta:
-        model = User
-        fields = ['first_name', 'last_name', 'email', 'password1', 'password2']
-
-
 class VoterRegistrationForm(forms.ModelForm):
     SELECT_GENDER = (
         (None, '-- Select your gender --'),
